Improved FileReader class
- fixed type conversion

+++ source/io/FileReader.cpp
@@ -3,7 +3,7 @@
  * Company:         Lynar Studios
  * E-Mail:          webmaster@lynarstudios.com
  * Created:         2020-08-17
- * Changed:         2020-08-18
+ * Changed:         2020-08-19
  *
  * */
 
@@ -24,7 +24,7 @@ ls_std::byte * ls_std::FileReader::read()
 {
   ls_std::byte* data;
   std::ifstream inputStream {this->file.getAbsoluteFilePath(), std::ifstream::binary};
-  int length = this->file.getSize();
+  int length = (int) this->file.getSize();
   data = new ls_std::byte[length];
   inputStream.read(data, length);
